Add message functionality
A user can now post messages and see other messages. This can only be
done by users that are logged in.
If the user is also a club member, the user can also see who wrote the
post and the time the post was created. Non club member can only see the
title and the content of the message.